requirements.txt
backcall==0.1.0
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
gitdb2==2.0.5
GitPython==2.1.11
h5py==2.8.0
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.0
pyparsing==2.3.1
python-dateutil==2.7.5
PyWavelets==1.0.1
PyYAML==3.13
scikit-image==0.14.2
scikit-learn==0.20.1
scikit-optimize==0.5.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
smmap2==2.0.5
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tqdm==4.28.1
traitlets==4.3.2
wcwidth==0.1.7
